refactor(contracts): extract contracts feature from players module
The `players/` feature had grown bloated, mixing player identity with
contract, contract-history, and player-transaction schemas. Move those
three schemas into a new `features/contracts/` feature so the
financial/roster-movement concepts have their own home.

No behavior change — service and repository logic still lives in
`players/` for now; a follow-up can extract the contract-specific
read/write logic into a dedicated contract service once the seam is
clearer.
