+++ server/features/players/players.repository.test.ts
@@ -5,9 +5,9 @@ import postgres from "postgres";
 import pino from "pino";
 import * as schema from "../../db/schema.ts";
 import { players } from "./player.schema.ts";
-import { contracts } from "./contract.schema.ts";
-import { contractHistory } from "./contract-history.schema.ts";
-import { playerTransactions } from "./player-transaction.schema.ts";
+import { contracts } from "../contracts/contract.schema.ts";
+import { contractHistory } from "../contracts/contract-history.schema.ts";
+import { playerTransactions } from "../contracts/player-transaction.schema.ts";
 import { playerAttributes } from "./attributes.schema.ts";
 import { playerDraftProfile } from "./player-draft-profile.schema.ts";
 import { playerSeasonStats } from "./player-career-log.schema.ts";
